fix PR notes
