+++ packages/database/package.json
@@ -28,7 +28,6 @@
     "@mikro-orm/postgresql": "^6.0.0-dev",
     "@mikro-orm/seeder": "^6.0.0-dev",
     "@mikro-orm/sql-highlighter": "^1.0.1",
-    "@hangar/shared": "*",
     "dotenv-flow": "^3.2.0",
     "mikro-orm": "^6.0.0-dev",
     "readline-sync": "^1.4.10"
