+++ packages/shared/package.json
@@ -15,8 +15,8 @@
     "test": "jest"
   },
   "dependencies": {
-    "@mikro-orm/core": "^5.7.14",
-    "@hangar/database": "*"
+    "@hangar/database": "*",
+    "dayjs": "^1.11.9"
   },
   "devDependencies": {
     "@types/jest": "^29.2.1",
