+++ packages/shared/src/index.ts
@@ -1,11 +1,13 @@
 export * from './config';
 export * from './types';
 export * from './utils';
-export { UserDTO,
-    ProjectDTO,
-    EventDTO,
-    PrizeDTO,
-    JudgeDTO,
-    JudgingVoteDTO,
-    AppConfigDTO
-} from '@hangar/database'
+export {
+  UserDTO,
+  ProjectDTO,
+  EventDTO,
+  PrizeDTO,
+  JudgeDTO,
+  JudgingVoteDTO,
+  AppConfigDTO,
+  AppConfigKey,
+} from '@hangar/database';
