+++ packages/shared/src/types/entities/AppConfig.ts
@@ -1,0 +1,12 @@
+import { AppConfigDTO } from '@hangar/database'
+import { Dayjs } from 'dayjs';
+
+export type AppConfig = Omit<AppConfigDTO, 'createdAt'|'updatedAt'> & {
+    createdAt: Dayjs
+    updatedAt: Dayjs
+}
+
+export type SerializedAppConfig = Omit<AppConfig, 'createdAt'|'updatedAt'> & {
+    createdAt: string
+    updatedAt: string
+}
