+++ packages/shared/src/types/entities/Event.ts
@@ -1,0 +1,16 @@
+import { EventDTO } from '@hangar/database'
+import { Dayjs } from 'dayjs';
+
+export type Event = Omit<EventDTO, 'createdAt'|'updatedAt'|'start'|'end'> & {
+    createdAt: Dayjs
+    updatedAt: Dayjs
+    start: Dayjs
+    end: Dayjs
+}
+
+export type SerializedEvent = Omit<Event, 'createdAt'|'updatedAt'|'start'|'end'> & {
+    createdAt: string
+    updatedAt: string
+    start: string
+    end: string
+}
