+++ packages/shared/src/types/entities/index.ts
@@ -1,0 +1,7 @@
+export * from './User'
+export * from './AppConfig'
+export * from './Judge'
+export * from './JudgingVotes'
+export * from './Project'
+export * from './Prize'
+export * from './Event'
