+++ packages/shared/src/types/index.ts
@@ -1,1 +1,2 @@
 export * from './ApiError';
+export * from './entities';
